apps/sample-traces-app/requirements.txt
flask>=2.3.0
psycopg2-binary>=2.9.0
redis>=4.5.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-http>=1.24.0
opentelemetry-instrumentation-flask>=0.45b0
opentelemetry-instrumentation-psycopg2>=0.45b0
opentelemetry-instrumentation-redis>=0.45b0
